03_Gold_Coast/Gold_Coast_Wide_Currently_For_Sale_AND_Recently_Sold/Ollama_Property_Analysis/requirements.txt
# Last Edit: 31/01/2026, Friday, 7:48 pm (Brisbane Time)
# Requirements for Ollama Property Analysis System

# MongoDB
pymongo>=4.6.1

# HTTP requests
requests>=2.31.0

# Image processing (flexible version to avoid conflicts)
Pillow>=10.2.0

# Environment variables
python-dotenv>=1.0.1

# Retry logic (if needed)
retry>=0.9.2
